add CardInfoManager
