+++ package.json
@@ -43,6 +43,7 @@
     "eslint-loader": "^1.6.1",
     "eslint-plugin-vue": "latest",
     "express": "^4.14.0",
+    "express-http-proxy": "^1.0.2",
     "extract-text-webpack-plugin": "2.0.0-beta.5",
     "file-loader": "^0.9.0",
     "hjson": "^1.7.4",
