+++ src/store.js
@@ -2,7 +2,8 @@ import Vue from 'vue'
 import Vuex from 'vuex'
 import _ from 'lodash'
 import { isLrigCard } from 'js/util'
-import Localize from 'js/Localize.js'
+import Localize from 'js/Localize'
+import { setCardInfo } from 'js/CardInfoManager'
 Vue.use(Vuex)
 
 /*
@@ -160,16 +161,7 @@ const mutations = {
       lang = 'zh_Hant'
     }
     state.localization.lang = lang
-    let langMaps = {
-      'zh_Hans': window.CardInfo_zh,
-      'zh_Hant': window.CardInfo_zh,
-      'en': window.CardInfo_en,
-      'jp': window.CardInfo_jp,
-      'ko': window.CardInfo_ko,
-      'ru': window.CardInfo_ru,
-      'it': window.CardInfo_it,
-    }
-    window.CardInfo = langMaps[lang]
+    setCardInfo(lang)
   },
 }
